finalizado função de venda em carrinho.jsp e vendaServlet

+++ src/br/com/gamestore/persistences/VendaManager.java
@@ -1,20 +1,19 @@
 package br.com.gamestore.persistences;
 
-import java.util.HashMap;
 import java.util.TreeMap;
 
 import br.com.gamestore.entidades.Venda;
 
 public class VendaManager {
 	
-	private static TreeMap<Integer, Venda> venda = new TreeMap<Integer, Venda>();
+	private static TreeMap<Double, Venda> venda = new TreeMap<Double, Venda>();
 	private static int cod = 0;
 
-	public static TreeMap<Integer, Venda> getVenda() {
+	public static TreeMap<Double, Venda> getVenda() {
 		return venda;
 	}
 
-	public static void setVenda(TreeMap<Integer, Venda> venda) {
+	public static void setVenda(TreeMap<Double, Venda> venda) {
 		VendaManager.venda = venda;
 	}
